+++ sshkey/tempdirprovider.go
@@ -1,0 +1,17 @@
+package sshkey
+
+import "github.com/bitrise-io/go-utils/pathutil"
+
+// TODO: pathutils / fs abstraction to go-utils
+
+// OsTempDirProvider ...
+type OsTempDirProvider struct{}
+
+// NewOsTempDirProvider ...
+func NewOsTempDirProvider() *OsTempDirProvider {
+	return &OsTempDirProvider{}
+}
+
+func (OsTempDirProvider) createTempDir(prefix string) (string, error) {
+	return pathutil.NormalizedOSTempDirPath(prefix)
+}
