+++ step/errors.go
@@ -1,0 +1,7 @@
+package step
+
+import "github.com/bitrise-io/go-steputils/step"
+
+func newStepError(tag string, err error, shortMsg string) *step.Error {
+	return step.NewError("activate-ssh-key", tag, err, shortMsg)
+}
